+++ schwabdev/client.py
@@ -7,7 +7,6 @@ import datetime
 import logging
 import asyncio
 import urllib.parse
-from typing import Any
 import threading
 import requests
 import aiohttp
@@ -20,7 +19,7 @@ class ClientBase:
 
     _base_api_url = "https://api.schwabapi.com"
 
-    def __init__(self, app_key, app_secret, callback_url="https://127.0.0.1", tokens_db="~/.schwabdev/tokens.db", timeout=10, call_on_auth=None):
+    def __init__(self, app_key, app_secret, callback_url="https://127.0.0.1", tokens_db="~/.schwabdev/tokens.db", encryption=None, timeout=10, call_on_auth=None):
         """
         Initialize a client to access the Schwab API.
 
@@ -28,7 +27,7 @@ class ClientBase:
             app_key (str): App key credential.
             app_secret (str): App secret credential.
             callback_url (str): URL for callback.
-            tokens_file (str): Path to tokens file.
+            tokens_db (str): Path to tokens file.
             timeout (int): Request timeout in seconds - how long to wait for a response.
             use_session (bool): Use a requests session for requests instead of creating a new session for each request.
             call_on_notify (function | None): Function to call when user needs to be notified (e.g. for input)
@@ -38,10 +37,10 @@ class ClientBase:
         if timeout <= 0:
             raise Exception("Timeout must be greater than 0 and is recommended to be 5 seconds or more.")
 
-        self.version = "Schwabdev 2.6.0"                                    # version of the client
+        self.version = "Schwabdev 3.0.0"                                    # version of the client
         self.timeout = timeout                                              # timeout to use in requests
         self.logger = logging.getLogger("Schwabdev")  # init the logger
-        self.tokens = Tokens(app_key, app_secret, callback_url, self.logger, tokens_db, call_on_auth)
+        self.tokens = Tokens(app_key, app_secret, callback_url, self.logger, tokens_db, encryption, call_on_auth)
         self.tokens.update_tokens()                                               # ensure tokens are up to date on init
 
     def _parse_params(self, params: dict):
@@ -122,7 +121,7 @@ class ClientBase:
 
 class Client(ClientBase):
 
-    def __init__(self, app_key, app_secret, callback_url="https://127.0.0.1", tokens_db="~/.schwabdev/tokens.db", timeout=10, call_on_auth=None):
+    def __init__(self, app_key:str, app_secret:str, callback_url:str="https://127.0.0.1", tokens_db: str="~/.schwabdev/tokens.db", encryption:str=None, timeout:int=10, call_on_auth:callable=None):
         """
         Initialize a client to access the Schwab API.
 
@@ -130,11 +129,11 @@ class Client(ClientBase):
             app_key (str): App key credential.
             app_secret (str): App secret credential.
             callback_url (str): URL for callback.
-            tokens_file (str): Path to tokens file.
+            tokens_db (str): Path to tokens file.
             timeout (int): Request timeout in seconds - how long to wait for a response.
             call_on_auth (function | None): Function to call for custom auth flow.
         """
-        super().__init__(app_key, app_secret, callback_url, tokens_db, timeout, call_on_auth)
+        super().__init__(app_key, app_secret, callback_url, tokens_db, encryption, timeout, call_on_auth)
 
         self._session = requests.Session()                                  # session to use in requests
         self._session.headers.update({'Authorization': f'Bearer {self.tokens.access_token}'})
@@ -160,11 +159,11 @@ class Client(ClientBase):
             return self._session.request(method, f'{self._base_api_url}{path}', timeout=self.timeout, **kwargs)
 
     def close(self):
-        with self._session_lock:
-            try:
+        try:
+            with self._session_lock:
                 self._session.close()
-            except Exception:
-                pass
+        except Exception as e:
+            self.logger.debug(f"{e} (Closed before full init?)")
 
     def __enter__(self):
         return self
@@ -401,9 +400,11 @@ class Client(ClientBase):
         return self._request("GET", f'/marketdata/v1/{urllib.parse.quote(symbol_id,safe="")}/quotes', 
                              params=self._parse_params({'fields': fields}))
 
-    def option_chains(self, symbol: str, contractType: str | None = None, strikeCount: Any = None, includeUnderlyingQuote: bool | None = None, strategy: str | None = None,
-               interval: Any = None, strike: Any = None, range: str | None = None, fromDate: datetime.datetime | str | None = None, toDate: datetime.datetime | str | None = None, volatility: Any = None, underlyingPrice: Any = None,
-               interestRate: Any = None, daysToExpiration: Any = None, expMonth: str | None = None, optionType: str | None = None, entitlement: str | None = None) -> requests.Response:
+    def option_chains(self, symbol: str, contractType: str | None = None, strikeCount: int | None = None, includeUnderlyingQuote: bool | None = None, 
+                      strategy: str | None = None, interval: str | None = None, strike: float | None = None, range: str | None = None, 
+                      fromDate: datetime.datetime | str | None = None, toDate: datetime.datetime | str | None = None, volatility: float | None = None, 
+                      underlyingPrice: float | None = None, interestRate: float | None = None, daysToExpiration: int | None = None, 
+                      expMonth: str | None = None, optionType: str | None = None, entitlement: str | None = None) -> requests.Response:
         """
         Get Option Chain including information on options contracts associated with each expiration for a ticker.
 
@@ -468,8 +469,9 @@ class Client(ClientBase):
         return self._request("GET", '/marketdata/v1/expirationchain', 
                              params=self._parse_params({'symbol': symbol}))
 
-    def price_history(self, symbol: str, periodType: str | None = None, period: str | None = None, frequencyType: str | None = None, frequency: Any = None, startDate: datetime.datetime | str | None = None,
-                      endDate: Any = None, needExtendedHoursData: bool | None = None, needPreviousClose: bool | None = None) -> requests.Response:
+    def price_history(self, symbol: str, periodType: str | None = None, period: str | None = None, frequencyType: str | None = None, 
+                      frequency: int | None = None, startDate: datetime.datetime | str | None = None, endDate: datetime.datetime | str | None = None, 
+                      needExtendedHoursData: bool | None = None, needPreviousClose: bool | None = None) -> requests.Response:
         """
         Get price history for a ticker
 
@@ -499,7 +501,7 @@ class Client(ClientBase):
                                                          'needPreviousClose': needPreviousClose}))
     
 
-    def movers(self, symbol: str, sort: str | None = None, frequency: Any = None) -> requests.Response:
+    def movers(self, symbol: str, sort: str | None = None, frequency: int | None = None) -> requests.Response:
         """
         Get movers in a specific index and direction
 
@@ -575,10 +577,10 @@ class Client(ClientBase):
 
 class ClientAsync(ClientBase):
 
-    def __init__(self, app_key, app_secret, callback_url="https://127.0.0.1", tokens_db="~/.schwabdev/tokens.db", timeout=10, call_on_auth=None, parsed:bool = False,):
+    def __init__(self, app_key:str, app_secret:str, callback_url:str="https://127.0.0.1", tokens_db: str="~/.schwabdev/tokens.db", encryption:str=None, timeout:int=10, call_on_auth:callable=None, parsed: bool = False,):
         if aiohttp is None:
             raise ImportError("aiohttp is required to use ClientAsync")
-        super().__init__(app_key, app_secret, callback_url, tokens_db, timeout, call_on_auth)
+        super().__init__(app_key, app_secret, callback_url, tokens_db, encryption, timeout, call_on_auth)
         self._parsed = parsed
         self._session = aiohttp.ClientSession(base_url=self._base_api_url,
                                               headers={'Authorization': f'Bearer {self.tokens.access_token}'}, 
@@ -616,7 +618,7 @@ class ClientAsync(ClientBase):
         retval = await self._task_group.__aexit__(exc_type, exc_val, exc_tb)
         return retval
     
-    async def _parse_response(self, response: aiohttp.ClientResponse, parsed: bool = None) -> aiohttp.ClientResponse | dict:
+    async def _parse_response(self, response: aiohttp.ClientResponse, parsed: bool | None = None) -> aiohttp.ClientResponse | dict:
         if (parsed is None and self._parsed) or (parsed is True):
             content_type = response.headers.get("Content-Type", "").lower()
             if content_type.startswith("application/json"):
@@ -634,7 +636,7 @@ class ClientAsync(ClientBase):
     Accounts and Trading Production
     """
 
-    async def linked_accounts(self, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def linked_accounts(self, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Account numbers in plain text cannot be used outside of headers or request/response bodies.
         As the first step consumers must invoke this service to retrieve the list of plain text/encrypted value pairs, and use encrypted account values for all subsequent calls for any accountNumber request.
@@ -647,7 +649,7 @@ class ClientAsync(ClientBase):
             parsed,
         )
 
-    async def account_details_all(self, fields: str = None, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def account_details_all(self, fields: str = None, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         All the linked account information for the user logged in. The balances on these accounts are displayed by default however the positions on these accounts will be displayed based on the "positions" flag.
 
@@ -665,7 +667,7 @@ class ClientAsync(ClientBase):
             parsed,
         )
         
-    async def account_details(self, accountHash: str, fields: str = None, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def account_details(self, accountHash: str, fields: str = None, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Specific account information with balances and positions. The balance information on these accounts is displayed by default but Positions will be returned based on the "positions" flag.
 
@@ -687,7 +689,7 @@ class ClientAsync(ClientBase):
     async def account_orders(self, accountHash: str, 
                              fromEnteredTime: datetime.datetime | str, 
                              toEnteredTime: datetime.datetime | str, 
-                             maxResults: int = None, status: str = None, parsed: bool = None) -> aiohttp.ClientResponse:
+                             maxResults: int = None, status: str = None, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         All orders for a specific account. Orders retrieved can be filtered based on input parameters below. Maximum date range is 1 year.
 
@@ -708,8 +710,8 @@ class ClientAsync(ClientBase):
                 params=self._parse_params(
                     {
                         'maxResults': maxResults,
-                        'fromEnteredTime': self._time_convert(fromEnteredTime, "8601"),
-                        'toEnteredTime': self._time_convert(toEnteredTime, "8601"),
+                        'fromEnteredTime': self._time_convert(fromEnteredTime, TimeFormat.ISO_8601),
+                        'toEnteredTime': self._time_convert(toEnteredTime, TimeFormat.ISO_8601),
                         'status': status,
                     }
                 ),
@@ -737,7 +739,7 @@ class ClientAsync(ClientBase):
             False, # must be raw to get headers
         )
 
-    async def order_details(self, accountHash: str, orderId: int | str, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def order_details(self, accountHash: str, orderId: int | str, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get a specific order by its ID, for a specific account
 
@@ -756,7 +758,7 @@ class ClientAsync(ClientBase):
         )
 
 
-    async def cancel_order(self, accountHash: str, orderId: int | str, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def cancel_order(self, accountHash: str, orderId: int | str, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Cancel a specific order by its ID, for a specific account
 
@@ -796,7 +798,7 @@ class ClientAsync(ClientBase):
 
     async def account_orders_all(self, fromEnteredTime: datetime.datetime | str, 
                                  toEnteredTime: datetime.datetime | str, 
-                                 maxResults: int = None, status: str = None, parsed: bool = None) -> aiohttp.ClientResponse:
+                                 maxResults: int = None, status: str = None, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get all orders for all accounts
 
@@ -816,8 +818,8 @@ class ClientAsync(ClientBase):
                 params=self._parse_params(
                     {
                         'maxResults': maxResults,
-                        'fromEnteredTime': self._time_convert(fromEnteredTime, "8601"),
-                        'toEnteredTime': self._time_convert(toEnteredTime, "8601"),
+                        'fromEnteredTime': self._time_convert(fromEnteredTime, TimeFormat.ISO_8601),
+                        'toEnteredTime': self._time_convert(toEnteredTime, TimeFormat.ISO_8601),
                         'status': status,
                     }
                 ),
@@ -826,7 +828,7 @@ class ClientAsync(ClientBase):
         )
 
 
-    async def preview_order(self, accountHash: str, order: dict, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def preview_order(self, accountHash: str, order: dict, parsed: bool | None = None) -> aiohttp.ClientResponse:
         return await self._parse_response(
             await self._session.post(
                 f'/trader/v1/accounts/{accountHash}/previewOrder',
@@ -839,7 +841,7 @@ class ClientAsync(ClientBase):
 
 
     async def transactions(self, accountHash: str, startDate: datetime.datetime | str, 
-                           endDate: datetime.datetime | str, types: str, symbol: str = None, parsed: bool = None) -> aiohttp.ClientResponse:
+                           endDate: datetime.datetime | str, types: str, symbol: str = None, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         All transactions for a specific account. Maximum number of transactions in response is 3000. Maximum date range is 1 year.
 
@@ -858,8 +860,8 @@ class ClientAsync(ClientBase):
                 f'/trader/v1/accounts/{accountHash}/transactions',
                 params=self._parse_params(
                     {
-                        'startDate': self._time_convert(startDate, "8601"),
-                        'endDate': self._time_convert(endDate, "8601"),
+                        'startDate': self._time_convert(startDate, TimeFormat.ISO_8601),
+                        'endDate': self._time_convert(endDate, TimeFormat.ISO_8601),
                         'symbol': symbol,
                         'types': types,
                     }
@@ -868,7 +870,7 @@ class ClientAsync(ClientBase):
             parsed,
         )
 
-    async def transaction_details(self, accountHash: str, transactionId: str | int, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def transaction_details(self, accountHash: str, transactionId: str | int, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get specific transaction information for a specific account
 
@@ -886,7 +888,7 @@ class ClientAsync(ClientBase):
             parsed,
         )
 
-    async def preferences(self, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def preferences(self, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get user preference information for the logged in user.
 
@@ -903,7 +905,7 @@ class ClientAsync(ClientBase):
     Market Data
     """
 
-    async def quotes(self, symbols : list[str] | str, fields: str = None, indicative: bool = False, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def quotes(self, symbols : list[str] | str, fields: str = None, indicative: bool = False, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get quotes for a list of tickers
 
@@ -929,7 +931,7 @@ class ClientAsync(ClientBase):
             parsed,
         )
 
-    async def quote(self, symbol_id: str, fields: str = None, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def quote(self, symbol_id: str, fields: str = None, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get quote for a single symbol
 
@@ -948,9 +950,11 @@ class ClientAsync(ClientBase):
             parsed,
         )
 
-    async def option_chains(self, symbol: str, contractType: str = None, strikeCount: any = None, includeUnderlyingQuote: bool = None, strategy: str = None,
-               interval: any = None, strike: any = None, range: str = None, fromDate: datetime.datetime | str = None, toDate: datetime.datetime | str = None, volatility: any = None, underlyingPrice: any = None,
-               interestRate: any = None, daysToExpiration: any = None, expMonth: str = None, optionType: str = None, entitlement: str = None, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def option_chains(self, symbol: str, contractType: str | None = None, strikeCount: int | None = None, includeUnderlyingQuote: bool | None = None, 
+                      strategy: str | None = None, interval: str | None = None, strike: float | None = None, range: str | None = None, 
+                      fromDate: datetime.datetime | str | None = None, toDate: datetime.datetime | str | None = None, volatility: float | None = None, 
+                      underlyingPrice: float | None = None, interestRate: float | None = None, daysToExpiration: int | None = None, 
+                      expMonth: str | None = None, optionType: str | None = None, entitlement: str | None = None, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get Option Chain including information on options contracts associated with each expiration for a ticker.
 
@@ -995,8 +999,8 @@ class ClientAsync(ClientBase):
                         'interval': interval,
                         'strike': strike,
                         'range': range,
-                        'fromDate': self._time_convert(fromDate, "YYYY-MM-DD"),
-                        'toDate': self._time_convert(toDate, "YYYY-MM-DD"),
+                        'fromDate': self._time_convert(fromDate, TimeFormat.YYYY_MM_DD),
+                        'toDate': self._time_convert(toDate, TimeFormat.YYYY_MM_DD),
                         'volatility': volatility,
                         'underlyingPrice': underlyingPrice,
                         'interestRate': interestRate,
@@ -1011,7 +1015,7 @@ class ClientAsync(ClientBase):
         )
 
 
-    async def option_expiration_chain(self, symbol: str, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def option_expiration_chain(self, symbol: str, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get an option expiration chain for a ticker
 
@@ -1029,8 +1033,9 @@ class ClientAsync(ClientBase):
             parsed,
         )
         
-    async def price_history(self, symbol: str, periodType: str = None, period: any = None, frequencyType: str = None, frequency: any = None, startDate: datetime.datetime | str = None,
-                        endDate: any = None, needExtendedHoursData: bool = None, needPreviousClose: bool = None, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def price_history(self, symbol: str, periodType: str | None = None, period: str | None = None, frequencyType: str | None = None, 
+                      frequency: int | None = None, startDate: datetime.datetime | str | None = None, endDate: datetime.datetime | str | None = None, 
+                      needExtendedHoursData: bool | None = None, needPreviousClose: bool | None = None, parsed: bool | None = None) -> aiohttp.ClientResponse:
             """
             Get price history for a ticker
 
@@ -1058,8 +1063,8 @@ class ClientAsync(ClientBase):
                         'period': period,
                         'frequencyType': frequencyType,
                         'frequency': frequency,
-                        'startDate': self._time_convert(startDate, 'epoch_ms'),
-                        'endDate': self._time_convert(endDate, 'epoch_ms'),
+                        'startDate': self._time_convert(startDate, TimeFormat.EPOCH_MS),
+                        'endDate': self._time_convert(endDate, TimeFormat.EPOCH_MS),
                         'needExtendedHoursData': self._handle_aiohttp_bool(needExtendedHoursData),
                         'needPreviousClose': self._handle_aiohttp_bool(needPreviousClose),
                     }
@@ -1068,7 +1073,7 @@ class ClientAsync(ClientBase):
             parsed,
         )
 
-    async def movers(self, symbol: str, sort: str = None, frequency: any = None, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def movers(self, symbol: str, sort: str = None, frequency: int | None = None, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get movers in a specific index and direction
 
@@ -1092,7 +1097,7 @@ class ClientAsync(ClientBase):
             parsed,
         )
 
-    async def market_hours(self, symbols: list[str], date: datetime.datetime | str = None, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def market_hours(self, symbols: list[str], date: datetime.datetime | str = None, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get Market Hours for dates in the future across different markets.
 
@@ -1109,14 +1114,14 @@ class ClientAsync(ClientBase):
                 params=self._parse_params(
                     {
                         'markets': symbols,
-                        'date': self._time_convert(date, 'YYYY-MM-DD'),
+                        'date': self._time_convert(date, TimeFormat.YYYY_MM_DD),
                     }
                 ),
             ),
             parsed,
         )
 
-    async def market_hour(self, market_id: str, date: datetime.datetime | str = None, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def market_hour(self, market_id: str, date: datetime.datetime | str = None, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get Market Hours for dates in the future for a single market.
 
@@ -1130,12 +1135,12 @@ class ClientAsync(ClientBase):
         return await self._parse_response(
             await self._session.get(
                 f'/marketdata/v1/markets/{market_id}',
-                params=self._parse_params({'date': self._time_convert(date, 'YYYY-MM-DD')}),
+                params=self._parse_params({'date': self._time_convert(date, TimeFormat.YYYY_MM_DD)}),
             ),
             parsed,
         )
 
-    async def instruments(self, symbol: str, projection: str, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def instruments(self, symbol: str, projection: str, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get instruments for a list of symbols
 
@@ -1154,7 +1159,7 @@ class ClientAsync(ClientBase):
             parsed,
         )
 
-    async def instrument_cusip(self, cusip_id: str | int, parsed: bool = None) -> aiohttp.ClientResponse:
+    async def instrument_cusip(self, cusip_id: str | int, parsed: bool | None = None) -> aiohttp.ClientResponse:
         """
         Get instrument for a single cusip